Replace user trait with a class which can be used to persist the object
